refactor: deduplicate some stuff
